+++ pytato/__init__.py
@@ -34,9 +34,6 @@ from pytato.array import (
 
         matmul, roll, transpose, stack, reshape, concatenate,
 
-        abs, sin, cos, tan, arcsin, arccos, arctan, sinh, cosh, tanh, exp, log,
-        log10, isnan, sqrt, conj, arctan2,
-
         maximum, minimum, where,
 
         full, zeros, ones,
@@ -47,12 +44,14 @@ from pytato.array import (
 
         logical_or, logical_and, logical_not,
 
-        real, imag,
-
         dot, vdot,
 
         )
 from pytato.reductions import sum, amax, amin, prod
+from pytato.cmath import (abs, sin, cos, tan, arcsin, arccos, arctan, sinh,
+                          cosh, tanh, exp, log, log10, isnan, sqrt, conj,
+                          arctan2, real, imag)
+
 
 from pytato.loopy import LoopyCall
 from pytato.target.loopy.codegen import generate_loopy
