+++ pytato/array.py
@@ -56,25 +56,6 @@ These functions generally follow the interface of the corresponding functions in
 .. autofunction:: transpose
 .. autofunction:: stack
 .. autofunction:: concatenate
-.. autofunction:: abs
-.. autofunction:: sqrt
-.. autofunction:: sin
-.. autofunction:: cos
-.. autofunction:: tan
-.. autofunction:: arcsin
-.. autofunction:: arccos
-.. autofunction:: arctan
-.. autofunction:: conj
-.. autofunction:: arctan2
-.. autofunction:: sinh
-.. autofunction:: cosh
-.. autofunction:: tanh
-.. autofunction:: exp
-.. autofunction:: log
-.. autofunction:: log10
-.. autofunction:: isnan
-.. autofunction:: real
-.. autofunction:: imag
 .. autofunction:: zeros
 .. autofunction:: ones
 .. autofunction:: full
@@ -94,6 +75,7 @@ These functions generally follow the interface of the corresponding functions in
 .. autofunction:: einsum
 .. autofunction:: dot
 .. autofunction:: vdot
+.. automodule:: pytato.cmath
 .. automodule:: pytato.reductions
 
 .. currentmodule:: pytato.array
@@ -600,8 +582,8 @@ class Array(Taggable):
         return pt.conj(self)
 
     def __abs__(self) -> Array:
-        # /!\ This refers to the abs() defined in this file, not the built-in.
-        return cast(Array, abs(self))
+        import pytato as pt
+        return cast(Array, pt.abs(self))
 
     def __pos__(self) -> Array:
         return self
@@ -1889,147 +1871,6 @@ def make_data_wrapper(data: DataInterface,
 # }}}
 
 
-# {{{ math functions
-
-def _apply_elem_wise_func(inputs: Tuple[ArrayOrScalar],
-                          func_name: str,
-                          ret_dtype: Optional[_dtype_any] = None
-                          ) -> ArrayOrScalar:
-    if all(isinstance(x, SCALAR_CLASSES) for x in inputs):
-        np_func = getattr(np, func_name)
-        return np_func(*inputs)  # type: ignore
-
-    if not inputs:
-        raise ValueError("at least one argument must be present")
-
-    shape = None
-
-    sym_args = []
-    bindings = {}
-    for index, inp in enumerate(inputs):
-        if isinstance(inp, Array):
-            if inp.dtype.kind not in ["f", "c"]:
-                raise ValueError("only floating-point or complex "
-                        "arguments supported")
-
-            if shape is None:
-                shape = inp.shape
-            elif inp.shape != shape:
-                # FIXME: merge this logic with arithmetic, so that broadcasting
-                # is implemented properly
-                raise NotImplementedError("broadcasting in function application")
-
-            if ret_dtype is None:
-                ret_dtype = inp.dtype
-
-            bindings[f"in_{index}"] = inp
-            sym_args.append(
-                    prim.Subscript(var(f"in_{index}"),
-                        tuple(var(f"_{i}") for i in range(len(shape)))))
-        else:
-            sym_args.append(inp)
-
-    assert shape is not None
-    assert ret_dtype is not None
-
-    return IndexLambda(
-            prim.Call(var(f"pytato.c99.{func_name}"), tuple(sym_args)),
-            shape, ret_dtype, bindings)
-
-
-def abs(x: Array) -> ArrayOrScalar:
-    if x.dtype.kind == "c":
-        result_dtype = np.empty(0, dtype=x.dtype).real.dtype
-    else:
-        result_dtype = x.dtype
-
-    return _apply_elem_wise_func((x,), "abs", ret_dtype=result_dtype)
-
-
-def sqrt(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "sqrt")
-
-
-def sin(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "sin")
-
-
-def cos(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "cos")
-
-
-def tan(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "tan")
-
-
-def arcsin(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "asin")
-
-
-def arccos(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "acos")
-
-
-def arctan(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "atan")
-
-
-def conj(x: Array) -> ArrayOrScalar:
-    if x.dtype.kind != "c":
-        return x
-    return _apply_elem_wise_func((x,), "conj")
-
-
-def arctan2(y: Array, x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((y, x), "atan2")  # type:ignore
-
-
-def sinh(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "sinh")
-
-
-def cosh(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "cosh")
-
-
-def tanh(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "tanh")
-
-
-def exp(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "exp")
-
-
-def log(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "log")
-
-
-def log10(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "log10")
-
-
-def isnan(x: Array) -> ArrayOrScalar:
-    return _apply_elem_wise_func((x,), "isnan", np.dtype(np.int32))
-
-
-def real(x: Array) -> ArrayOrScalar:
-    if x.dtype.kind == "c":
-        result_dtype = np.empty(0, dtype=x.dtype).real.dtype
-    else:
-        return x
-    return _apply_elem_wise_func((x,), "real", ret_dtype=result_dtype)
-
-
-def imag(x: Array) -> ArrayOrScalar:
-    if x.dtype.kind == "c":
-        result_dtype = np.empty(0, dtype=x.dtype).real.dtype
-    else:
-        return zeros(x.shape, dtype=x.dtype)
-    return _apply_elem_wise_func((x,), "imag", ret_dtype=result_dtype)
-
-# }}}
-
-
 # {{{ full
 
 def full(shape: ConvertibleToShape, fill_value: ScalarType,
@@ -2246,6 +2087,7 @@ def maximum(x1: ArrayOrScalar, x2: ArrayOrScalar) -> ArrayOrScalar:
     array-like objects that could be broadcasted together. NaNs are propagated.
     """
     # https://github.com/python/mypy/issues/3186
+    from pytato.cmath import isnan
     return where(logical_or(isnan(x1), isnan(x2)), np.NaN,  # type: ignore
                  where(greater(x1, x2), x1, x2))
 
@@ -2256,6 +2098,7 @@ def minimum(x1: ArrayOrScalar, x2: ArrayOrScalar) -> ArrayOrScalar:
     array-like objects that could be broadcasted together. NaNs are propagated.
     """
     # https://github.com/python/mypy/issues/3186
+    from pytato.cmath import isnan
     return where(logical_or(isnan(x1), isnan(x2)), np.NaN,  # type: ignore
                  where(less(x1, x2), x1, x2))
